# openapi/httpfs.yaml
openapi: 3.0.3
...
info:
  title: SFTPGo HTTPFs
  description: |
    SFTPGo can use custom storage backend implementations compliant with the API defined here.
    HTTPFs is a work in progress and makes no API stability promises.
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    url: 'https://www.gnu.org/licenses/agpl-3.0.en.html'
servers:
- url: /v1
security:
- ApiKeyAuth: []
- BasicAuth: []
paths:
...
  /truncate/{name}:
    parameters:
      - name: name
        in: path
        description: object name
        required: true
        schema:
          type: string
      - name: size
        in: query
        required: true
        description: 'new file size in bytes'
        schema:
          type: integer
          format: int64
    patch:
      tags:
        - fs
      summary: Changes the size of the named file
      operationId: truncate
      responses:
        200:
          $ref: '#/components/responses/OKResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
        501:
          $ref: '#/components/responses/NotImplemented'
        default:
          $ref: '#/components/responses/DefaultResponse'
  /readdir/{name}:
    parameters:
      - name: name
        in: path
        description: object name
        required: true
        schema:
          type: string
    get:
      tags:
        - fs
      summary: Reads the named directory and returns the contents
      operationId: readdir
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileInfo'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
        501:
          $ref: '#/components/responses/NotImplemented'
        default:
          $ref: '#/components/responses/DefaultResponse'
...
components:
  responses:
    OKResponse:
      description: successful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    NotImplemented:
      description: Not Implemented
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    DefaultResponse:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          description: 'message, can be empty'
        error:
          type: string
          description: error description if any
    FileInfo:
      type: object
      properties:
        name:
          type: string
          description: base name of the file
        size:
          type: integer
          format: int64
          description: length in bytes for regular files; system-dependent for others
        mode:
          type: integer
          description: |
            File mode and permission bits. More details here: https://golang.org/pkg/io/fs/#FileMode.
            Let's see some examples:
            - for a directory mode&2147483648 != 0
            - for a symlink mode&134217728 != 0
            - for a regular file mode&2401763328 == 0
        last_modified:
          type: string
          format: date-time
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY